+++ test/coverage/test_combinations.py
@@ -66,13 +66,13 @@ def assert_not_conform(values: list, schema: dict):
 
 @pytest.fixture
 def ctx():
-    return CoverageContext(location="query", is_required=True, custom_formats=get_default_format_strategies())
+    return CoverageContext(location="body", is_required=True, custom_formats=get_default_format_strategies())
 
 
 @pytest.fixture
 def pctx():
     return CoverageContext(
-        location="query",
+        location="body",
         generation_modes=[GenerationMode.POSITIVE],
         is_required=True,
         custom_formats=get_default_format_strategies(),
@@ -82,7 +82,7 @@ def pctx():
 @pytest.fixture
 def nctx():
     return CoverageContext(
-        location="query",
+        location="body",
         generation_modes=[GenerationMode.NEGATIVE],
         is_required=True,
         custom_formats=get_default_format_strategies(),