+++ CHANGELOG.md
@@ -2,6 +2,10 @@
 
 ## [Unreleased](https://github.com/schemathesis/schemathesis/compare/v4.0.8...HEAD) - TBD
 
+### :wrench: Changed
+
+- Do not display API probing result in the test phases summary.
+
 ## [4.0.8](https://github.com/schemathesis/schemathesis/compare/v4.0.7...v4.0.8) - 2025-07-12
 
 ### :bug: Fixed