+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 
 - Do not display API probing result in the test phases summary.
 - Do not display API probing errors in CLI output. Treat errors as a missing capability.
+- Display cURL code samples on more network errors. [#2940](https://github.com/schemathesis/schemathesis/issues/2940) 
 
 ## [4.0.8](https://github.com/schemathesis/schemathesis/compare/v4.0.7...v4.0.8) - 2025-07-12